+++ src/commands/index.ts
@@ -1,5 +1,8 @@
 // import pxToEm from "./pxToEm";
 import emToPx from "./emToPx";
 import pxToEm from "./pxToEm";
+import remToPx from "./remToPx";
+import pxToRem from "./pxToRem";
+import changeRootFont from "./changeRootFont";
 
-export { emToPx, pxToEm };
+export { emToPx, pxToEm, remToPx, pxToRem, changeRootFont };