+++ src/extension.ts
@@ -1,14 +1,19 @@
-import * as vscode from "vscode";
-import { pxToEm, emToPx } from "./commands/index";
+import { commands, ExtensionContext } from "vscode";
+import {
+  pxToEm,
+  emToPx,
+  pxToRem,
+  remToPx,
+  changeRootFont,
+} from "./commands/index";
 
-// TODO:
-// - refactor callback function in registerCommand()
-// -- refactor to use hooks (consideration)
-
-export function activate(context: vscode.ExtensionContext) {
+export function activate(context: ExtensionContext) {
   context.subscriptions.push(
-    vscode.commands.registerCommand("px-to-em.pxToEm", () => pxToEm()),
-    vscode.commands.registerCommand("px-to-em.emToPx", () => emToPx())
+    commands.registerCommand("px-to-em.pxToEm", () => pxToEm()),
+    commands.registerCommand("px-to-em.emToPx", () => emToPx()),
+    commands.registerCommand("px-to-em.pxToRem", () => pxToRem()),
+    commands.registerCommand("px-to-em.remToPx", () => remToPx()),
+    commands.registerCommand("px-to-em.changeRootFont", () => changeRootFont())
   );
 }